requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.1
django-ninja==0.19.1
django-on-heroku==1.1.2
google-auth==2.10.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
gspread-dataframe==3.3.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
numpy==1.23.1
oauthlib==3.2.0
outcome==1.2.0
pandas==1.4.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
whitenoise==6.2.0
wsproto==1.1.0
